+++ R/hdbscan.R
@@ -70,7 +70,7 @@
 #' explicitly computed (see cluster tree in Chaudhuri et al, 2010).
 #' @param gen_simplified_tree logical; should the simplified hierarchy be
 #' explicitly computed (see Campello et al, 2013).
-#' @oaram verbose report progress.
+#' @param verbose report progress.
 #' @param ...  additional arguments are passed on.
 #' @param scale integer; used to scale condensed tree based on the graphics
 #' device. Lower scale results in wider trees.