+++ R/AAA_dbscan-package.R
@@ -1,6 +1,6 @@
 #' @title `r packageDescription("dbscan")$Package`: `r packageDescription("dbscan")$Title`
 #'
-#' @description `r packageDescription("dbscan")$Description`
+#' @description `r sub("<[^>]+>", "", packageDescription("dbscan")$Description)`
 #'
 #' @section Key functions:
 #' - Clustering: [dbscan()], [hdbscan()], [optics()], [jpclust()], [sNNclust()]
@@ -8,6 +8,10 @@
 #' - Nearest Neighbors: [kNN()], [frNN()], [sNN()]
 #'
 #' @author Michael Hahsler and Matthew Piekenbrock
+#'
+#' @references
+#' Hahsler M, Piekenbrock M, Doran D (2019). dbscan: Fast Density-Based Clustering with R. Journal of Statistical Software, 91(1), 1-30. \doi{10.18637/jss.v091.i01}
+#'
 #' @docType package
 #' @name dbscan-package
 #'