+++ R/dbscan.R
@@ -149,7 +149,7 @@
 #' ## Find suitable DBSCAN parameters:
 #' ## 1. We use minPts = dim + 1 = 5 for iris. A larger value can also be used.
 #' ## 2. We inspect the k-NN distance plot for k = minPts - 1 = 4
-#' kNNdistplot(iris, k = 5 - 1)
+#' kNNdistplot(iris, minPts = 5)
 #'
 #' ## Noise seems to start around a 4-NN distance of .7
 #' abline(h=.7, col = "red", lty = 2)
@@ -393,4 +393,4 @@ print.dbscan_fast <- function(x, ...) {
 #' @rdname dbscan
 #' @export
 is.corepoint <- function(x, eps, minPts = 5, ...)
-  sapply(frNN(x, eps = 0.5, ...)$id, length) >= (minPts - 1)
+  sapply(frNN(x, eps = eps, ...)$id, length) >= (minPts - 1)