+++ NEWS.md
@@ -1,12 +1,16 @@
-# dbscan 1.1-10-1 (2022-xx-xx)
+# dbscan 1.1-11 (2022-10-21)
 
 ## New Features
+* kNNdistplot gained parameter minPts.
 * dbscan now retains information on distance method and border points.
 * HDBSCAN now supports long vectors to work with larger distance matrices. 
 * conversion from dist to kNN and frNN is now more memory efficient. It does no longer 
   coerce the dist object into a matrix of double the size, but extract the distances directly
   from the dist object.
 * Better description of how predict uses only Euclidean distances and more error checking.
+
+## Bugfix
+* is.corepoint() now uses the correct epsilon value (reported by Eng Aun)
 
 # dbscan 1.1-10 (2022-01-14)